Add first working version of the TYPO3 DebugBar
This handles the injection of the assets and the DebugBar code to the
frontend and the possibility to configure the extension.

The DebugBar is enabled by default and will displayed as soon as a
user logs in the backend as an Admin.

It is possible to disable the DebugBar in the extension settings.
Its further possible to en/disable the data collectors.

The following collectors are available:

* PhpInfoCollector: Shows information about PHP. Currently the version.
* MessagesCollector: Add possibility to add Messages from anywhere in the application to the DebugBar.
* TimeCollector: Add rendering time.
* MemoryCollector: Add memory usage.
* ExceptionsCollector: Add the possibility to show Exceptions in the DebugBar.
* AuthCollector: Shows information about the logged-in BE user. Currently only the username.
* InfoCollector: Shows basically the same data as the Info pane of the Adminpanel.
* MySqliCollector: Shows all fired SQL statements.
* Typo3Collector: Shows information about the TYPO3 instance like locale, environment and version.