Add Exception and Measuring logging support
To implement this, I had to change the usage of the hooks. Now the Typo3DebugBar is
registered right after the BE user was initialized. The instance of the BE User is passed
to the constructor of the DebugBar to check if the its enabled later on.

After registration the DebugBar is booted on the same hook. This is because we need to have it all setup up
on a very early point so that it is possible to add Exceptions and Timers.

The tricky part is, that this BE user is a "maybe BE user" because later in the Bootstrap process it might
get invalidated if the user has no access to the very page. For now its close enough but should be take into
consideration how to proceed here. Maybe open up a PR which adds a custom hook right after the final BE user is available.