Extend the JavaScriptRenderer and move all asset related code there
In the prior versions I manually copied the default styles and JavaScript code from the PHP DebugBar library to

This patch streamlines the handling of all assets. While in prior versions, the default assets where copied manually from the PHP DebugBar
to this extension the new version will use the API of PHP DebugBar to use there default assets and combine them with custom assets.

Additionally the font files for FontAwesome are now included in the style.css file.

This changes reducing the package size and its easier to update later on because the extension overwrites only what needed.

The code of the various CSS files will be combined into one file and stored to the upload folder of the extension. Same happens to the JavaScript files.