Render the params into the SQL query
In the database pane the statements were shown as Prepared Statements with the placeholders instead of
values. The values are shown below each statement in a hidden table. The user has to click on the table
to display it. Some users want to see the values directly inside the statements.

I added an option to the extension settings to enable/disable this feature. Additionally I refactored the
maintaining of different connections. Now the extension uses the `Default` connection of the first connection within
the ConnectionPool. Later on I will implement a setting to define a specific connection.

The commit also removes the css and js widgets for the database pane because PHP Debugbar already ships it and I did
not change it. Take it from there, less to maintain for me.

Closes #8